Bundle Asio on RHEL 8
Asio is not currently available for RHEL 8, but it appears that Fast-DDS
is able to bundle it. Rather than introduce an Asio package to the
bootstrap repository just for this package, we can flip this flag to
bundle it.

The CMake logic in Fast-DDS expects the sources to be in a git
repository, and uses submodules to download the bundled sources. When
the package is bloomed, the submodules are initialized and included in
the tarball, but the sources are no longer part of a git repository.
Since we know that the directory has already been populated, we need to
suppress the submodule initialization and update step.

Signed-off-by: Scott K Logan <logans@cottsay.net>